Rewrite the OG membership classes to use a base and extend. Removed a duplicate class function as we can use now use the standard OgMembershipUser and provide a role_name yaml array to filter on. Added OG Membership Content migration and test migration files.
